Add Quick events support in Quickspace
This adds a new UI to Quickspace when events are available such as Now
Playing, Alarms, or whatever we configure as an event. This logic is
inspired by Google's event infrastructure in Smartspace but completely
written from scratch to be housed in Quickspace.

Initial events are:

- Device intro: Shows on first boot and setup
- Ambient Play (Now Playing): Shows what song is playing and tapping
the song will launch the results in Google search

Change-Id: I2742c13043b816c54bc2cefdfd415f52f2e76d3f

spezi77 edits:
- Fix package name for DoubleShadowTextView
- Update strings for ROM intro
- Update weather client package

Bootleggers edit:
- Let's disable the Now Playing code, might recycle it later.
- Migrate to OmniJaws

Signed-off-by: spezi77 <spezi7713@gmx.net>
Change-Id: I3e523bfabb773626bc128110306edcf3f9fd2d59
Signed-off-by: pri0818 <priyanshusinghal0818@gmail.com>