Set finishedToApp = false when rotating in Overview
The issue is that we always assume finishedToApp = true when recents animation is cancelled and use that to set FLAG_IN_APP. However, we don't return to app in overview after the device is rotated. Ideally we want to get signal from onRecentsAnimationCancelled, but at this point in the release it's probably ok to use launcher state to determine it.

Fixes: 227565959
Test: manual
Change-Id: I6be800aff8108b382ca3947cd1d0a45c9e9571eb
Signed-off-by: strwbrry143 <alexfinhart@gmail.com>
Signed-off-by: pri0818 <priyanshusinghal0818@gmail.com>